Still closer
The next piece of the puzzle is that when the JSON data contains two following
transactions:

    REMOVAL OF OPTION DUE TO ASSIGNMENT
    OPTION ASSIGNMENT

This is counted by TD as a single order, where the P/L from closing the option
position is transferred into the cost basis of the equity open, if such an
opening event occurs. Otherwise, it's just straight PL from the expiry.

Further, for transaction of type:

    REMOVAL OF OPTION DUE TO EXPIRATION

The price must be set to zero, not null.